+++ BusinessLogic/Services/UserFileSevice.cs
@@ -21,17 +21,29 @@ namespace BusinessLogic.Services
         {
             var userfile = await _repositoryWrapper.UserFile
                 .FindByCondition(x => x.UserFileId == id);
+            if (userfile is null || userfile.Count == 0)
+            {
+                throw new ArgumentNullException("Not found");
+            }
             return userfile.First();
         }
 
         public async Task Create(UserFile model)
         {
+            if (model == null)
+            {
+                throw new ArgumentNullException(nameof(model));
+            }
             _repositoryWrapper.UserFile.Create(model);
             _repositoryWrapper.Save();
         }
 
         public async Task Update(UserFile model)
         {
+            if (model == null)
+            {
+                throw new ArgumentNullException(nameof(model));
+            }
             _repositoryWrapper.UserFile.Update(model);
             _repositoryWrapper.Save();
         }
@@ -40,7 +52,10 @@ namespace BusinessLogic.Services
         {
             var userfile = await _repositoryWrapper.UserFile
                 .FindByCondition(x => x.UserFileId == id);
-
+            if (userfile is null || userfile.Count == 0)
+            {
+                throw new ArgumentNullException("Not found");
+            }
             _repositoryWrapper.UserFile.Delete(userfile.First());
             _repositoryWrapper.Save();
         }
